requirements.txt
numpy==1.18.5
streamlit==0.62.1
sklearn==0.0
scikit-learn== 0.23.1
pandas== 1.0.5
pillow== 7.2.0
Pillow @ file:///C:/ci/pillow_1594304973959/work
scikit-learn @ file:///C:/ci/scikit-learn_1592863447244/work
pandas @ file:///C:/ci/pandas_1592841744841/work